tests, fixes to TimeUtils

+++ VirtualChannels/ITimeUtils.cs
@@ -2,7 +2,7 @@
 {
     public interface ITimeUtils<TTime>
     {
-        TTime IndexToTime(TTime time0, int index, int rateInMilliHz);
+        TTime IndexToTime(TTime time0, long index, int rateInMilliHz);
         long TimeToIndex(TTime time0, TTime currentTime, int rateInMilliHz);
         TTime Zero();
     }
